Remove inverted and undocumented whitelist.
The behavior of whitelist is inverted. Any jobs
present in whitelist are skipped. This is the
exact opposite of the generally accepted meaning
of whitelist.

whitelist is redundant. The current code pattern
and behavior of whitelist exactly duplicate
jobs.deny. If the behavior of whitelist were
flipped and, thereby, aligned with the normal
historical meaning, the code pattern and behavior
would exactly duplicate jobs.allow. In either
case, this feature is redundant.

whitelist is undocumented. Conversely, both
jobs.allow and jobs.deny have been documented for
the recorded lifetime of this git project.

It seems sensible to simply remove this inverted,
undocumented feature.