Prioritize existence check to improve readability and perf.
run-parts already skips nonexistent files which have made
it into the loop list. It does so for all current cases:
the flagless case, the --list case, and the --test case.
This change maintains all of these preexisting behaviors.

It makes sense to prioritize the nonexistent case above the
other checks in this loop. Put another way, if the file
doesn't exist, we really don't need to test if it's a
directory, has a skipped suffix/extension, or is/isn't in
jobs.{allow,deny}.

This change both 1) makes the code read easier and
2) improves performance for certain corner cases.

Regarding readability, this change reuses the "or continue"
pattern already present in this script. It also removes one
level of nesting in the densest section of code.

Casual performance testing against dangling symlinks (which
match the for loop glob, but fail existence tests) was done
on Ryzen 5700G.

[Commands]
```
$ mkdir 1{00,000,0000,00000}_dangling_symlinks
$ for i in 1{00,000,0000,00000}; do
>   for j in `seq 1 "$i"`; do
>     ln -s null "$i"_dangling_symlinks/dangling_symlink_$j;
>   done;
> done
$ for i in 1{00,000,0000,00000}_dangling_symlinks; do
>   echo "$i";
>   time ./run-parts "$i";
> done
```

[Results]
run-parts (@9e74f2d):
                          real      user      sys
   100_dangling_symlinks  0m0.008s  0m0.003s  0m0.004s
  1000_dangling_symlinks  0m0.043s  0m0.032s  0m0.012s
 10000_dangling_symlinks  0m0.403s  0m0.340s  0m0.064s
100000_dangling_symlinks  0m3.880s  0m3.315s  0m0.575s

run-parts (patched):
                          real      user      sys
   100_dangling_symlinks  0m0.003s  0m0.000s  0m0.003s
  1000_dangling_symlinks  0m0.008s  0m0.003s  0m0.006s
 10000_dangling_symlinks  0m0.066s  0m0.049s  0m0.018s
100000_dangling_symlinks  0m0.659s  0m0.480s  0m0.189s

Very casually, we see a 500-600% performance improvement.

These numbers of dangling symlinks are, hopefully, a purely
synthetic case. However, given how many systems exist which
run-parts many times a day, this performance improvement
has a reasonable chance of being material at a global and
long-term scale.